move admin server into it's own servce
