+++ src/services/AdminServer/AdminServer.test.ts
@@ -5,7 +5,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 import { AdminServer } from "./AdminServer";
-import { MCServer } from "./services/MCServer/MCServer";
+import { MCServer } from "../MCServer/MCServer";
 
 const adminServer = new AdminServer(new MCServer());
